🚚 Move Card.jsx to partials folder
